feat: 걷기 task google maps 추가, GPS 기록 및 권한 취득

+++ lib/models/sliderController.dart
@@ -1,4 +1,5 @@
 class SliderController {
-  double sliderValue;
+  int sliderValue;
+
   SliderController(this.sliderValue);
 }
